Refactored routes for WHP Selection Tool, moved json data files to controllers folder instead of routes folder, minor amendments for typos
