+++ app/views/index.html
@@ -34,7 +34,7 @@
           {% set whpSelectionToolWorkcoachScreens = [{
             workTitle : 'Gatekeeper',
             links: [{ 
-              link : 'latest/whp-annual-profile', title : 'Record WHP places available'
+              link : 'latest/gatekeeper/profile', title : 'Record WHP places available'
             }, {
           link : 'latest/whp-weekly-profile', title : 'Profile WHP places across the year'
           }, {
