+++ app/views/latest/whp-annual-profile.html
@@ -18,7 +18,7 @@
               <option value="Select">Select</option>
               <option value="Bradford">Bradford</option>
               <option value="Derby">Derby</option>
-              <option value="Liverpool" selected>Liverpool</option>
+              <option value="Liverpool" selected>Mercia</option>
               <option value="Manchester">Manchester</option>
               <option value="Portsmouth">Portsmouth</option>
               <option value="Southampton">Southampton</option>
@@ -90,7 +90,7 @@
         <div class="column-third">
           <div class="form-group">
             <label class="form-label" for="whp-end-date">WHP end date</label>
-            <input class="form-control form-control-1-4" type="text" id="whp-end-date" value="01/08/17">
+            <input class="form-control form-control-1-4" type="text" id="whp-end-date" value="01/08/18">
           </div>
         </div>
       </div>
