+++ app/views/latest/whp-selection-report.html
@@ -57,7 +57,7 @@
                   <td>Y</td>
                   <td></td>
                   <td>
-                    <input name='random-allocation-1', id="random-allocation-1", type="checkbox", checked="checked")></input>
+                    <input name="random-allocation-1", id="random-allocation-1", type="checkbox", checked="checked")></input>
                   </td>
                 </tr>
                 <tr>
@@ -67,7 +67,7 @@
                   <td>Y</td>
                   <td></td>
                   <td>
-                    <input name='random-allocation-2', id="random-allocation-2", type="checkbox")></input>
+                    <input name="random-allocation-2", id="random-allocation-2", type="checkbox")></input>
                   </td>
                 </tr>
                 <tr>
@@ -77,7 +77,7 @@
                   <td></td>
                   <td>Y</td>
                   <td>
-                    <input name='random-allocation-3', id="random-allocation-3", type="checkbox", checked="checked")></input>
+                    <input name="random-allocation-3", id="random-allocation-3", type="checkbox", checked="checked")></input>
                   </td>
                 </tr>
                 <tr>
@@ -87,7 +87,7 @@
                   <td></td>
                   <td></td>
                   <td>
-                    <input name='random-allocation-4', id="random-allocation-4", type="checkbox", checked="checked")></input>
+                    <input name="random-allocation-4", id="random-allocation-4", type="checkbox", checked="checked")></input>
                   </td>
                 </tr>
                 <tr>
@@ -97,7 +97,7 @@
                   <td>" "</td>
                   <td>" "</td>
                   <td>
-                    <input name='random-allocation-0', id="random-allocation-0", type="checkbox")></input>
+                    <input name="random-allocation-0", id="random-allocation-0", type="checkbox")></input>
                   </td>
                 </tr>
                 <tr>
@@ -107,7 +107,7 @@
                   <td></td>
                   <td></td>
                   <td>
-                    <input name='random-allocation-70', id="random-allocation-70", type="checkbox")></input>
+                    <input name="random-allocation-70", id="random-allocation-70", type="checkbox")></input>
                   </td>
                 </tr>
               </tbody>
