+++ server.js
@@ -201,7 +201,7 @@ if (promoMode === 'true') {
     res.send('User-agent: *\nAllow: /')
   })
 } else {
-  // Disallow search index idexing
+  // Disallow search index indexing
   app.use(function (req, res, next) {
     // Setting headers stops pages being indexed even if indexed pages link to them.
     res.setHeader('X-Robots-Tag', 'noindex')
